requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
tabulate>=0.9
requests>=2.28
python-dotenv>=1.0
